Created a React App
In this we first deleted the app built in files as we were going to create everything from scratch. After gathering all the assets that were going to be used into the system, we simply added them to the src folder. Then we created the app component to test the working of this website.